requirements.txt
numpy==1.26.4
opencv-python==4.10.0.84
mediapipe==0.10.9
PySide6==6.7.2
python-vlc==3.0.21203
pyserial==3.5
